requirements.txt
asgiref==3.11.0
Django==5.2.9
django-environ==0.12.0
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
